Make a bootstrap by keeping a copy of err.asm and sbl.asm
Building x64 spitbol requires a usable spitbol binary.

Keep a copy of the 3 files that are created by running lex.sbl, asm.sbl
and err.sbl.  This will allow building spitbol without having a spitbol
binary.

Note that sbl.lex isn't really needed by the bootstrap process, but
including it in the bootstrap directory for checking the bootstrap
code.

To bootstrap from scratch:

$ make bootsbl
$ make BASEBOL=./bootsbl spitbol

Run the sanity-check command to ensure the newly built spitbol "works"
$ ./sanity-check

Look for errors in the diff.

To verify that the bootstrap files are still in sync, do a make spitbol
then
$ make checkboot

Again look for errors in the diff.

If there are changes to lex.sbl, asm.sbm or err.sbl rebuild the
bootstrap directory
$ make makeboot